error : 클라이언트 컴포넌트 에러 예시화
