+++ README.md
@@ -12,6 +12,6 @@
 
 # 서버 컴포넌트 사용
 - [x] 데이터를 호출해야 하는 경우, 서버 컴포넌트로 분류한다.
-  - [ ] 서버 컴포넌트는 클라이언트 컴포넌트에서 호출이 불가능하다.
+  - [ ] 서버 컴포넌트는 클라이언트 컴포넌트에서 호출(import)이 불가능하다. -> {children}으로 호출한다. 이는, 비동기 렌더링을 적용하기 위함이다.
   - [ ] useState, useEffect와 같은 생명 주기, 상태를 가질 수 없다. 즉, 1번 렌더링 된다.
 - [ ] 클라이언트 컴포넌트와 서버 컴포넌트를 엄격히 분리한다.
