+++ app/folder/Content.tsx
@@ -1,0 +1,7 @@
+const Content = () => {
+  return (
+    <div>Content</div>
+  )
+}
+
+export default Content
