+++ app/folder/page.tsx
@@ -1,14 +1,16 @@
+import Content from './Content';
 import Gnb from '@/components/gnb/Gnb';
 import AddLink from '@/components/addLink/AddLink';
 import Footer from '@/components/footer/Footer';
 
 const FolderPage = () => {
   return (
-    <div>
+    <>
       <Gnb isFixed={false}/>
       <AddLink />
+      <Content />
       <Footer />
-    </div>
+    </>
   )
 }
