+++ components/gnb/Gnb.tsx
@@ -1,8 +1,12 @@
+'use client'
+
+import { useState } from 'react'
 import Link from 'next/link'
 import Image from 'next/image'
 
-import { getUser } from '@/apis/api'
 import logoImg from '@/public/logo.svg'
+
+import GnbProfile from './GnbProfile'
 
 import './Gnb.css'
 
@@ -10,19 +14,8 @@ interface GnbProps {
   isFixed: boolean;
 }
 
-const getProfileData = async () => {
-  try {
-    const { data } = await getUser();
-
-    return data;
-  } catch {
-    throw new Error('error')
-  }
-}
-
-// server component : 상태와 부수 효과, 생명주기를 가질 수 없음.
-const Gnb = async ({ isFixed }: GnbProps) => {
-  const [user] = await getProfileData();
+const Gnb = ({ isFixed }: GnbProps) => {
+  const [count, setCount] = useState(0)
 
   const gnbStyle: React.CSSProperties = {
     position: isFixed ? 'sticky' : 'static',
@@ -36,11 +29,9 @@ const Gnb = async ({ isFixed }: GnbProps) => {
         <Link href="/">
           <Image className='logo' src={logoImg} alt="홈으로 연결된 Linkbrary 로고" />
         </Link>
-        <div className='account'>
-          <img className="account-img" src={user.image_source} alt={`${user.email} 프로필`} />
-          <span className='account-email'>{user.email}</span>
-        </div>
+        <GnbProfile />
       </div>
+      <button onClick={() => setCount(count + 1)}>{count}</button>
     </nav> 
   )
 } // img 태그 사용 이유 : Image는 최적화가 가능한 대신, public에 존재해야 한다. 즉, 서버에서 다운로드한 이미지는 사용 불가.
